Updated models, logic, handlers, dotnet build local script

+++ src/DShop.Services.Products/Handlers/DeleteProductHandler.cs
@@ -1,5 +1,6 @@
 ﻿using DShop.Common.Handlers;
 using DShop.Common.RabbitMq;
+using DShop.Common.Types;
 using DShop.Services.Products.Messages.Commands;
 using DShop.Services.Products.Messages.Events;
 using DShop.Services.Products.Repositories;
@@ -27,12 +28,22 @@ namespace DShop.Services.Products.Handlers
             => await _handler
                 .Handle(async () =>
                 {
+                    if (!await _productsRepository.ExistsAsync(command.Id))
+                    {   
+                        throw new DShopException("product_not_found",
+                            $"Product with id: '{command.Id}' was not found.");
+                    }
                     await _productsRepository.DeleteAsync(command.Id);
                 })
-                .OnSuccess(async () =>
-                {
-                    await _busPublisher.PublishAsync(new ProductDeleted(command.Id), context);
-                })
+                .OnSuccess(async () => await _busPublisher.PublishAsync(
+                    new ProductDeleted(command.Id), context)
+                )
+                .OnCustomError(async ex => await _busPublisher.PublishAsync(
+                    new DeleteProductRejected(command.Id, ex.Message, ex.Code), context)
+                )
+                .OnError(async ex => await _busPublisher.PublishAsync(
+                    new DeleteProductRejected(command.Id, ex.Message, "delete_product_failed"), context)
+                )
                 .ExecuteAsync();
     }
 }
